move manifests to context
