+++ src/App.tsx
@@ -5,21 +5,24 @@ import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Main from "./components/Main";
 import { BrowserRouter as Router, Route, Redirect } from "react-router-dom";
+import { ManifestsProvider } from "./contexts/ManifestsContext";
 
 function App() {
   return (
     <div className="App">
       <ThemeProvider>
-        <Router>
-          <Route path="/search">
-            <Header />
-            <Main />
-          </Route>
-          <Route path="*">
-            <Redirect to="/search" />
-          </Route>
-        </Router>
-        <Footer />
+        <ManifestsProvider>
+          <Router>
+            <Route path="/search">
+              <Header />
+              <Main />
+            </Route>
+            <Route path="*">
+              <Redirect to="/search" />
+            </Route>
+          </Router>
+          <Footer />
+        </ManifestsProvider>
       </ThemeProvider>
     </div>
   );
