+++ src/types/Props.ts
@@ -1,0 +1,3 @@
+export default interface Props {
+  children?: React.ReactNode;
+}
